correct MovieDataSource
