+++ app/src/main/java/com/example/compmoviedb/di/NetworkCommonModule.kt
@@ -1,10 +1,6 @@
 package com.example.compmoviedb.di
 
-import com.example.compmoviedb.data.repository.MoviesRepositoryImpl
-import com.example.compmoviedb.data.storage.MoviesStorage
-import com.example.compmoviedb.data.storage.themoviedb.MovieDBApiInterface
-import com.example.compmoviedb.data.storage.themoviedb.TheMovieDBMoviesStorageImpl
-import com.example.compmoviedb.domain.repository.MoviesRepository
+import com.example.compmoviedb.data.dataSource.themoviedb.MovieDBApiInterface
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
