+++ app/src/main/java/com/example/compmoviedb/di/NetworkModule.kt
@@ -1,9 +1,9 @@
 package com.example.compmoviedb.di
 
 import com.example.compmoviedb.data.repository.MoviesRepositoryImpl
-import com.example.compmoviedb.data.storage.MoviesStorage
-import com.example.compmoviedb.data.storage.themoviedb.MovieDBApiInterface
-import com.example.compmoviedb.data.storage.themoviedb.TheMovieDBMoviesStorageImpl
+import com.example.compmoviedb.data.dataSource.MoviesDataSource
+import com.example.compmoviedb.data.dataSource.themoviedb.MovieDBApiInterface
+import com.example.compmoviedb.data.dataSource.themoviedb.TheMovieDBMoviesDataSource
 import com.example.compmoviedb.domain.repository.MoviesRepository
 import dagger.Module
 import dagger.Provides
@@ -17,11 +17,11 @@ object NetworkModule {
 
     @Provides
     @Singleton
-    fun providesMoviesStorage(movieDBApiInterface: MovieDBApiInterface): MoviesStorage =
-        TheMovieDBMoviesStorageImpl(movieDBApiInterface = movieDBApiInterface)
+    fun providesMoviesStorage(movieDBApiInterface: MovieDBApiInterface): MoviesDataSource =
+        TheMovieDBMoviesDataSource(movieDBApiInterface = movieDBApiInterface)
 
     @Provides
     @Singleton
-    fun providesMoviesRepository(moviesStorage: MoviesStorage): MoviesRepository =
+    fun providesMoviesRepository(moviesStorage: MoviesDataSource): MoviesRepository =
         MoviesRepositoryImpl(moviesStorage = moviesStorage)
 }
