+++ data/src/main/java/com/example/compmoviedb/data/dataSource/MoviesDataSource.kt
@@ -1,13 +1,13 @@
-package com.example.compmoviedb.data.storage
+package com.example.compmoviedb.data.dataSource
 
-import com.example.compmoviedb.data.storage.models.movieactors.ListActorsMovieEntity
-import com.example.compmoviedb.data.storage.models.moviedetails.MovieDetailsEntity
-import com.example.compmoviedb.data.storage.models.moviespopular.ListMoviesPopularEntity
-import com.example.compmoviedb.data.storage.models.movievideo.MovieVideoEntity
+import com.example.compmoviedb.data.dataSource.models.movieactors.ListActorsMovieEntity
+import com.example.compmoviedb.data.dataSource.models.moviedetails.MovieDetailsEntity
+import com.example.compmoviedb.data.dataSource.models.moviespopular.ListMoviesPopularEntity
+import com.example.compmoviedb.data.dataSource.models.movievideo.MovieVideoEntity
 import com.example.compmoviedb.domain.models.Response
 import kotlinx.coroutines.flow.Flow
 
-interface MoviesStorage {
+interface MoviesDataSource {
     suspend fun getMoviesPopular(): Flow<Response<ListMoviesPopularEntity>>
 
     suspend fun getMovieDetailsById(movieId: Int): Flow<Response<MovieDetailsEntity>>
