+++ data/src/main/java/com/example/compmoviedb/data/dataSource/models/movieactors/Cast.kt
@@ -1,14 +1,15 @@
-package com.example.compmoviedb.data.storage.models.movieactors
+package com.example.compmoviedb.data.dataSource.models.movieactors
 
-data class Crew(
+data class Cast(
     val adult: Boolean?,
+    val cast_id: Int?,
+    val character: String?,
     val credit_id: String?,
-    val department: String?,
     val gender: Int?,
     val id: Int?,
-    val job: String?,
     val known_for_department: String?,
     val name: String?,
+    val order: Int?,
     val original_name: String?,
     val popularity: Double?,
     val profile_path: String?
