+++ data/src/main/java/com/example/compmoviedb/data/dataSource/models/moviedetails/Genre.kt
@@ -1,0 +1,6 @@
+package com.example.compmoviedb.data.dataSource.models.moviedetails
+
+data class Genre(
+    val id: Int?,
+    val name: String?
+)
