+++ data/src/main/java/com/example/compmoviedb/data/dataSource/models/moviedetails/MovieDetailsEntity.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.data.storage.models.moviedetails
+package com.example.compmoviedb.data.dataSource.models.moviedetails
 
 data class MovieDetailsEntity(
     val adult: Boolean?,
