+++ data/src/main/java/com/example/compmoviedb/data/dataSource/models/moviedetails/SpokenLanguage.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.data.storage.models.moviedetails
+package com.example.compmoviedb.data.dataSource.models.moviedetails
 
 data class SpokenLanguage(
     val iso_639_1: String?,
