+++ data/src/main/java/com/example/compmoviedb/data/dataSource/models/moviespopular/ListMoviesPopularEntity.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.data.storage.models.moviespopular
+package com.example.compmoviedb.data.dataSource.models.moviespopular
 
 data class ListMoviesPopularEntity(
     val page: Int,
