+++ data/src/main/java/com/example/compmoviedb/data/dataSource/models/movievideo/MovieVideoEntity.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.data.storage.models.movievideo
+package com.example.compmoviedb.data.dataSource.models.movievideo
 
 data class MovieVideoEntity(
     val id: Int?,
