+++ data/src/main/java/com/example/compmoviedb/data/dataSource/themoviedb/MovieDBConstants.kt
@@ -1,0 +1,10 @@
+package com.example.compmoviedb.data.dataSource.themoviedb
+
+object MovieDBConstants {
+    const val MOVIE_DB_KEY = "e57c71791b41b1b6d48678746aa69e44"
+}
+
+enum class LanguageCodes() {
+    RU,
+    EN
+}
