+++ data/src/main/java/com/example/compmoviedb/data/mappers/moviedetails/ListGenreEntityMapper.kt
@@ -1,7 +1,7 @@
 package com.example.compmoviedb.data.mappers.moviedetails
 
 import com.example.compmoviedb.data.mappers.MapperData
-import com.example.compmoviedb.data.storage.models.moviedetails.Genre
+import com.example.compmoviedb.data.dataSource.models.moviedetails.Genre
 import com.example.compmoviedb.domain.models.moviedetails.GenreD
 
 class ListGenreEntityMapper : MapperData<List<Genre>, List<GenreD>> {
