+++ data/src/main/java/com/example/compmoviedb/data/mappers/moviedetails/ListSpokenLanguageEntityMapper.kt
@@ -1,7 +1,7 @@
 package com.example.compmoviedb.data.mappers.moviedetails
 
 import com.example.compmoviedb.data.mappers.MapperData
-import com.example.compmoviedb.data.storage.models.moviedetails.SpokenLanguage
+import com.example.compmoviedb.data.dataSource.models.moviedetails.SpokenLanguage
 import com.example.compmoviedb.domain.models.moviedetails.SpokenLanguageD
 
 class ListSpokenLanguageEntityMapper : MapperData<List<SpokenLanguage>, List<SpokenLanguageD>> {
