+++ data/src/main/java/com/example/compmoviedb/data/mappers/movievideo/MovieVideoEntityMapper.kt
@@ -1,7 +1,7 @@
 package com.example.compmoviedb.data.mappers.movievideo
 
 import com.example.compmoviedb.data.mappers.MapperData
-import com.example.compmoviedb.data.storage.models.movievideo.MovieVideoEntity
+import com.example.compmoviedb.data.dataSource.models.movievideo.MovieVideoEntity
 import com.example.compmoviedb.domain.models.movievideo.MovieVideoD
 
 class MovieVideoEntityMapper : MapperData<MovieVideoEntity, MovieVideoD> {
