+++ data/src/main/java/com/example/compmoviedb/data/mappers/movievideo/ResultListMovieVideoMapper.kt
@@ -1,7 +1,7 @@
 package com.example.compmoviedb.data.mappers.movievideo
 
 import com.example.compmoviedb.data.mappers.MapperData
-import com.example.compmoviedb.data.storage.models.movievideo.Result
+import com.example.compmoviedb.data.dataSource.models.movievideo.Result
 import com.example.compmoviedb.domain.models.movievideo.ResultMovieVideoDetailsD
 
 class ResultListMovieVideoMapper : MapperData<List<Result>, List<ResultMovieVideoDetailsD>> {
