--- data/src/main/java/com/example/compmoviedb/data/storage/themoviedb/MovieDBConstants.kt
@@ -1,6 +1,0 @@
-package com.example.compmoviedb.data.storage.themoviedb
-
-object MovieDBConstants {
-    const val MOVIE_DB_KEY = "e57c71791b41b1b6d48678746aa69e44"
-    const val LANGUAGE_RU = "ru"
-}
